Split GameFieldViewModel to separate files
