+++ src/JoinRpg.WebPortal.Models/FieldSetup/GameFieldListViewModel.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+
+namespace JoinRpg.Web.Models.FieldSetup
+{
+    public class GameFieldListViewModel
+    {
+        public IEnumerable<GameFieldEditViewModel> Items { get; }
+
+        public FieldNavigationModel Navigation { get; }
+
+        public GameFieldListViewModel(FieldNavigationModel navigation, IEnumerable<GameFieldEditViewModel> fields)
+        {
+            Navigation = navigation;
+            Items = fields;
+        }
+    }
+}
