+++ src/JoinRpg.WebPortal.Models/FieldSetup/IFieldNavigationAware.cs
@@ -1,0 +1,10 @@
+namespace JoinRpg.Web.Models.FieldSetup
+{
+    public interface IFieldNavigationAware : IProjectIdAware
+    {
+        FieldNavigationModel Navigation { get; }
+        new int ProjectId { get; set; }
+
+        void SetNavigation(FieldNavigationModel fieldNavigation);
+    }
+}
